Create Codsoft_task1
A calculator is a machine that allows people to do math operations more easily. For example, most calculators will add, subtract, multiply, and divide. Some also do square roots, and more complex calculators can help with calculus and draw function graphs. In this code, I just showed a sample of a calculator and how it works.