Create Codsoft_task2
A password generator is a software tool that creates random or customized passwords for users. It helps users create stronger passwords that provide greater security for a given type of access. 
Password generators help those who have to constantly come up with new passwords to ensure authorized access to programs and to manage a large number of passwords for identity and access management.